requirements.txt
asgiref==3.6.0
canvas==0.3
certifi==2023.7.22
charset-normalizer==3.3.0
crispy-bootstrap4==2022.1
Django==4.2
django-appconf==1.0.5
django-auditlog==2.3.0
django-autocomplete-light==3.9.7
django-bootstrap-form==3.4
django-bootstrap-modal-forms==3.0.2
django-ckeditor==6.5.1
django-crispy-forms==2.0
django-htmx==1.14.0
django-js-asset==2.0.0
django-materializecss-form==1.1.17
django-select2==8.1.2
django-widget-tweaks==1.4.12
idna==3.4
mysqlclient==2.1.1
numpy==1.24.2
Pillow==10.0.1
python-dateutil==2.8.2
reportlab==4.0.5
requests==2.31.0
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
urllib3==2.0.6
XlsxWriter==3.1.4
